requirements.txt
asgiref==3.8.1
babel==2.17.0
Django==5.1.7
django-money==3.5.3
djangorestframework==3.16.0
psycopg2==2.9.10
py-moneyed==3.0
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
